requirements.txt
bandit==1.8.6
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.3
rich==14.1.0
SQLAlchemy==2.0.43
stevedore==5.5.0
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
